Added ogg audio fixture and test
Also lowered bitrate on videos to 300k to decrease total repo size
